+++ mmcv/ops/csrc/parrots/mlu/nms_mlu.cpp
@@ -1,70 +1,19 @@
 #include <parrots/compute/aten.hpp>
 #include <parrots_mlu_helper.hpp>
-#include <type_traits>
 
 using namespace parrots;
-
-#define USE_CPU_NMS
 
 template <typename T>
 void nms_parrots(T& ctx, const SSElement& attr,
                  const OperatorBase::in_list_t& ins,
                  OperatorBase::out_list_t& outs) {}
 
+#define USE_CPU_NMS
+
 #ifdef USE_CPU_NMS
 
 using at::Tensor;
-
-Tensor nms_cpu(Tensor boxes, Tensor scores, float iou_threshold, int offset) {
-  if (boxes.numel() == 0) {
-    return at::empty({0}, boxes.options().dtype(at::kLong));
-  }
-  auto x1_t = boxes.select(1, 0).contiguous();
-  auto y1_t = boxes.select(1, 1).contiguous();
-  auto x2_t = boxes.select(1, 2).contiguous();
-  auto y2_t = boxes.select(1, 3).contiguous();
-
-  Tensor areas_t = (x2_t - x1_t + offset) * (y2_t - y1_t + offset);
-
-  auto order_t = std::get<1>(scores.sort(0, /* descending=*/true));
-
-  auto nboxes = boxes.size(0);
-  Tensor select_t = at::ones({nboxes}, boxes.options().dtype(at::kBool));
-
-  auto select = select_t.data_ptr<bool>();
-  auto order = order_t.data_ptr<int64_t>();
-  auto x1 = x1_t.data_ptr<float>();
-  auto y1 = y1_t.data_ptr<float>();
-  auto x2 = x2_t.data_ptr<float>();
-  auto y2 = y2_t.data_ptr<float>();
-  auto areas = areas_t.data_ptr<float>();
-
-  for (int64_t _i = 0; _i < nboxes; _i++) {
-    if (select[_i] == false) continue;
-    auto i = order[_i];
-    auto ix1 = x1[i];
-    auto iy1 = y1[i];
-    auto ix2 = x2[i];
-    auto iy2 = y2[i];
-    auto iarea = areas[i];
-
-    for (int64_t _j = _i + 1; _j < nboxes; _j++) {
-      if (select[_j] == false) continue;
-      auto j = order[_j];
-      auto xx1 = std::max(ix1, x1[j]);
-      auto yy1 = std::max(iy1, y1[j]);
-      auto xx2 = std::min(ix2, x2[j]);
-      auto yy2 = std::min(iy2, y2[j]);
-
-      auto w = std::max(0.f, xx2 - xx1 + offset);
-      auto h = std::max(0.f, yy2 - yy1 + offset);
-      auto inter = w * h;
-      auto ovr = inter / (iarea + areas[j] - inter);
-      if (ovr > iou_threshold) select[_j] = false;
-    }
-  }
-  return order_t.masked_select(select_t);
-}
+Tensor nms_cpu(Tensor boxes, Tensor scores, float iou_threshold, int offset);
 
 template <>
 void nms_parrots<HostContext>(HostContext& ctx, const SSElement& attr,
@@ -193,7 +142,9 @@ PARROTS_EXTENSION_REGISTER(nms)
     .attr("offset")
     .input(2)
     .output(1)
+#ifdef USE_CPU_NMS
     .apply(nms_parrots<HostContext>)
+#endif
 #ifdef PARROTS_USE_CAMB
     .apply(nms_parrots<CambContext>)
 #endif
